+++ tests/test_subcommands.py
@@ -5,7 +5,7 @@ Test subcommands.
 import logging
 
 from uwgroups.scripts.main import main
-from __init__ import TestCaseSuppressOutput, TestBase
+from .__init__ import TestCaseSuppressOutput, TestBase
 
 log = logging.getLogger(__name__)
 
@@ -17,6 +17,3 @@ class TestTemplate(TestCaseSuppressOutput, TestBase):
 
     def testExit02(self):
         self.assertRaises(SystemExit, main, ['-h'])
-
-    # def test01(self):
-    #     main(['template', 'infile', 'outfile', '--monkey-type', 'macaque'])