+++ uwgroups/__init__.py
@@ -1,5 +1,11 @@
 """
 CLI for the UW Groups API
+
+Use the environment variable GWS_HOST to define the API host:
+
+PROD: groups.uw.edu (default)
+DEV: dev.groups.uw.edu
+EVAL: eval.groups.uw.edu
 """
 
 import glob
@@ -34,5 +40,5 @@ def package_data(fname, pattern=None):
 try:
     with open(path.join(path.dirname(__file__), 'data', 'ver')) as f:
         __version__ = f.read().strip().replace('-', '+', 1).replace('-', '.')
-except Exception, e:
+except Exception:
     __version__ = ''