+++ uwgroups/subcommands/members.py
@@ -1,7 +1,9 @@
-"""Create a group
+"""List group members
 """
 
 import logging
+import pprint
+import sys
 
 from uwgroups.api import UWGroups
 from uwgroups.subcommands import find_credentials
@@ -11,12 +13,14 @@ log = logging.getLogger(__name__)
 
 def build_parser(parser):
     parser.add_argument('group_name', help="name of a group")
-    parser.add_argument('--exchange', choices=['yes', 'no'], default='yes')
 
 
 def action(args):
     certfile, keyfile = find_credentials(args)
     with UWGroups(certfile, keyfile) as conn:
-        conn.create_group(args.group_name)
-        conn.set_affiliate(
-            args.group_name, service='exchange', active=args.exchange == 'yes')
+        if conn.group_exists(args.group_name):
+            members = conn.get_members(args.group_name)
+            for m in sorted(members):
+                print(m)
+        else:
+            sys.exit(f'group "{args.group_name}" does not exist')